requirements_cloud.txt
streamlit
langchain
langchain-community
langchain-core
langchain-text-splitters
deep-translator
langdetect
pypdf
python-docx
requests